requirements.txt
streamlit
google-generativeai
Pillow
websocket-client